+++ pkg/front_end/lib/src/fasta/kernel/body_builder.dart
@@ -4027,8 +4027,9 @@ class BodyBuilder extends StackListenerImpl
     }
     push(new InternalRecordLiteral(
         positional, named, namedElements, originalElementOrder,
-        isConst: constKeyword != null ||
-            constantContext == ConstantContext.inferred));
+        isConst:
+            constKeyword != null || constantContext == ConstantContext.inferred)
+      ..fileOffset = token.offset);
   }
 
   void buildLiteralSet(List<TypeBuilder>? typeArguments, Token? constKeyword,